Add initial health check if statement

+++ p2f_client/datasets.py
@@ -1,5 +1,6 @@
 # Local libraries
 from p2f_pydantic.datasets import Datasets
+from .conn import health_check
 # Third Party Libraries
 import requests
 from furl import furl
@@ -17,16 +18,18 @@ class datasets:
         self.upload_queue.append(dataset)
     def upload_datasets(self):
         uploaded_datasets = []
-        for dataset in self.upload_queue:
-            r = requests.post(self.dataset_url,
-                              data=dataset.model_dump_json(exclude_unset=True))
-            uploaded_datasets.append(Datasets(**r.json()))
-        # self.uploaded_datasets = uploaded_datasets
-        return uploaded_datasets
+        if health_check(self.base_url):
+            for dataset in self.upload_queue:
+                r = requests.post(self.dataset_url,
+                                data=dataset.model_dump_json(exclude_unset=True))
+                uploaded_datasets.append(Datasets(**r.json()))
+            # self.uploaded_datasets = uploaded_datasets
+            return uploaded_datasets
     def upload_dataset(self, dataset: Datasets):
-        r = requests.post(self.dataset_url,
-                          data=dataset.model_dump_json(exclude_unset=True))
-        return Datasets(**r.json())
+        if health_check(self.base_url):
+            r = requests.post(self.dataset_url,
+                            data=dataset.model_dump_json(exclude_unset=True))
+            return Datasets(**r.json())
     def list_remote_datasets(self, 
                              is_new_p2f: Optional[bool]=None,
                              is_sub_dataset: Optional[bool]=None,
@@ -42,13 +45,16 @@ class datasets:
         if doi is not None:
             list_url.args["doi"] = doi
             # data["doi"] = doi
-        r = requests.get(list_url)
-        # self.datasets = [Datasets(**x) for x in r.json()]
-        return [Datasets(**x) for x in r.json()]
+        if health_check(self.base_url):
+            r = requests.get(list_url)
+            # self.datasets = [Datasets(**x) for x in r.json()]
+            return [Datasets(**x) for x in r.json()]
     def get_remote_dataset(self, dataset_id):
         get_url = self.dataset_url / str(dataset_id)
-        r = requests.get(get_url)
-        return Datasets(**r.json())
+        if health_check(self.base_url):
+            r = requests.get(get_url)
+            return Datasets(**r.json())
     def delete_remote_dataset(self, dataset_id):
         delete_url = self.dataset_url / str(dataset_id)
-        r = requests.delete(delete_url)
+        if health_check(self.base_url):
+            r = requests.delete(delete_url)
