+++ p2f_client/harm_data_types.py
@@ -1,5 +1,6 @@
 # Local libraries
 from p2f_pydantic.harm_data_types import harm_data_type as Harm_data_type
+from .conn import health_check
 # Third Party Libraries
 import requests
 from furl import furl
@@ -16,13 +17,15 @@ class harm_data_type:
     def add_harm_data_type(self, new_data_type: Harm_data_type):
         self.harm_data_types_queue.append(new_data_type)
     def upload_data_types(self):
-        for datatype in self.harm_data_types_queue:
-            r = requests.post(self.hdt_url,
-                              data=datatype.model_dump_json(exclude_unset=True))
+        if health_check(self.base_url):
+            for datatype in self.harm_data_types_queue:
+                r = requests.post(self.hdt_url,
+                                data=datatype.model_dump_json(exclude_unset=True))
     def upload_data_type(self, new_data_type: Harm_data_type):
-        r = requests.post(self.hdt_url,
-                          data=new_data_type.model_dump_json(exclude_unset=True))
-        return Harm_data_type(**r.json())
+        if health_check(self.base_url):
+            r = requests.post(self.hdt_url,
+                            data=new_data_type.model_dump_json(exclude_unset=True))
+            return Harm_data_type(**r.json())
     def list_data_types(self, 
                         measure: Optional[str]=None,
                         unit_of_measure: Optional[str]=None,
@@ -33,11 +36,14 @@ class harm_data_type:
                   "method": method, 
                   "dataset_id": dataset_id}
         params = {x:y for x, y in params.items() if y != None}
-        r = requests.get(self.hdt_url,
-                         params=params)
-        return [Harm_data_type(**x) for x in r.json()]
+        if health_check(self.base_url):
+            r = requests.get(self.hdt_url,
+                            params=params)
+            return [Harm_data_type(**x) for x in r.json()]
     def get_data_type(self, datatype_id: UUID):
-        r = requests.get(self.hdt_url / datatype_id)
-        return Harm_data_type(**r.json())
+        if health_check(self.base_url):
+            r = requests.get(self.hdt_url / datatype_id)
+            return Harm_data_type(**r.json())
     def delete_data_type(self, datatype_id: UUID):
-        r = requests.delete(self.hdt_url / datatype_id)
+        if health_check(self.base_url):
+            r = requests.delete(self.hdt_url / datatype_id)
